Extend the EpochInfo window in the Alonzo era.
The `EpochInfo` provided to the ledger by consensus can throw an error
(or return the Left of an Either) when the slot it is asked to translate
lies outside of its predictable window. However, at present this
function has a bug: when consensus doesn't know about any future epochs,
it considers its predictable interval to be unbounded. This may then
change when a new era is added to the codebase, resulting in old
transactions not validating.

The intention is to remove this bug in consensus, such that the
`EpochInfo` is appropriately bounded in all eras. Unfortunately, since
some transactions with TTLs outside of the usual window have been
allowed on-chain, we have to disable this check in the Alonzo era. In
order to do this, we linearly extend the `EpochInfo` based on the
current and previous slot, which we are guaranteed to be able to
translate, since otherwise we could not be processing this slot.