Restrict TTL to lie within the stability windown.
The `EpochInfo` provided to the ledger by consensus can throw an error
(or return the Left of an Either) when the slot it is asked to translate
lies outside of its predictable window. However, when consensus doesn't
know about any future epochs, it considers its predictable interval to
be unbounded. This makes sense in the context of its original reason for
creation, which is validating blocks, but has the unfortunate
consequence that relying on it in the ledger will allow unbounded
translation of TTL in the final era. This may then change when a new era
is added to the codebase.

This PR adds additional bounds into the ledger to ensure that, in any
case, the TTL should be bounded by the stability window.

Note that we cannot entirely avoid relying on the `EpochInfo` (e.g.
turning the `Left` hand side into an `error`), since it is possible for
the `EpochInfo` to _restrict_ the valid translation interval (e.g. if it
knows about a new epoch where time translation is only possible for half
the number of slots). This potential nondeterminism should hopefully not
be an issue, since it should only depend on the _slot number_, which is
fixed for a given transaction.

Unfortunately, since some transactions with TTLs outside of the usual
window have been allowed on-chain, we have to disable this check in the
Alonzo era. In the future (once we have transitioned to Babbage), we may
wish to replace this check with specific exclusions for the transactions
which exhibit this.