Require the caller of `checkLeaderValue` to provide bounds.
The input to `checkLeaderValue` is now not guaranteed to be the output
of a VRF evaluation. Consequently, we require the input to explicitly
specify its bounds.

We only provide one way to construct a 'BoundedNatural', which will
throw an error if the bound is violated. This is most appropriate for
our use case, since we "know" the bounds statically, and a violation
indicates a programmer error.