Ensure pure `EpochInfo` is not overused.
The `EpochInfo` in `Globals` sits in `Either Text`, but we allow it to
be exposed in a pure `Identity` monad for cases where the information
solicited is guaranteed to be available. However, since this is the
"default" option, at some points it has been misused. As such, we make
three changes:

1. The names have been reversed: `epochInfo` will now give the `Either
   Text` version, whereas `epochInfoPure` will give the `Identity`
   version. The latter has some additional notes describing when it's
   valid to use these.
2. Not all places that previously used `epochInfo` are switched to
   `epochInfoPure`. In particular, we ensure that places that use time
   translation must use the full `EpochInfo (Either Text)`.
3. We rationalise the logic around `txInfo`. We now take the `EpochInfo
   (Either Text)` explicitly and fail inside the same `Left` case as all
   other failures.

Co-authored-by: Alexey Kuleshevich <alexey.kuleshevich@iohk.io>