fix dangerous use of Arary.!
findAndCount (used in a CLI helper function) was using Array.! in a
place where is was not guaranteed to succeed. We now check the bounds.

Additionally, one of the errors had a misleading name.
RedeemerNotNeeded was actually accouting for a case when a redeemer
pointed to a script hash which we could not resolve.