Refactor `BabbageFeatures`
Each testcase defines the transaction to be validated, along with the
outputs that should exist in utxo at the time of the validation.
The invokes one of the three functions: `testExpectSuccessValid`,
`testExpectSuccessInvalid` or `testExpectFailure`, depending on what we
want to test.
These functions, based on the given transaction and outputs, construct
the initial and actual utxo and check it against the expected one.