Change "Follow the Satoshi" to title case
To a reader unfamiliar with the "Follow the Satoshi" algorithm, this
previously looked like a typo as if the author couldn't decide between
"in contrast to following the Satoshi algorithm" and "in contrast to
using the Satoshi algorithm". By rendering "Follow the Satoshi" in title
case it becomes clear to the reader that "Follow the Satoshi" is the name
of an algorithm and not a typo.