Avoid ghc bug that causes an infinite loop
A more restricitve constraint with a redundant associated type family
constraint causes GHC to generate code that goes into an infinite loop.
This depicts iteslf in an OOM when compiled with `-O0` and non
terminating computaton when compiled with optimizations.

Here is the bug report on the GHC issue tracker:
https://gitlab.haskell.org/ghc/ghc/-/issues/21973