Update nix shell to prebuild packages and make them available to ghc-pkg
so that running cabal from within the shell doesn't have to (which
sometimes fails).

Update plutus version and adjust to API changes

Update cardano-base to a version compatible with plutus

Allow newer version of `hashable`

since plutus brings in version 1.4.1.0 but our dependency Unique is
restricting it to 1.4.

Regenerate PlutusScript using plutus-preprocessor

Fix Plutus protocol version in test

which was wrong, but only started failing with the new plutus version