Add `getScriptsNeeded` to the `EraUTxO`:
This approach allows us to move `mintedTxBodyF` into ShelleyMA era,
where it belongs. It also unifies the spec and code, where spec uses the
same function `scriptsNeeded` in all eras. In order to achieve this we
need a special type family `ScriptsNeeded`, since eras do not agree in
the return type of the function.