Bump to use latest released cardano-base and cardano-prelude packages (#3080)
* Bump to use latest released cardano-base and cardano-prelude packages

The release branch is using an unreleased commit of `cardano-base` which
contains the SECP primitives but is before the `heapwords` split. This
makes it incompatible with either of the released versions:
- The previous version doesn't have the SECP primitives
- The next version has the heapwords split

In order to make the release branch usable via CHaP, I think the
simplest thing is to backport the changes for the newer `cardano-base`
packages, i.e. the `heapwords` split. This is a bit tedious but not too
bad overall.

Afterwards I'll release another version from this branch.

* ormolise

* Bump minor version

1.0.0 -> 1.1.1
1.3.0 -> 1.4.1

* Move to CHaP

I think we might have a few over-specified constraints here.

Co-authored-by: Moritz Angermann <moritz.angermann@gmail.com>