WIP: Absolutely _MUTILATE_ the TICKF rule
Esgen and I poked and prodded the rule with pd' sharing.

We do have evidence that it saves time, but:

- I have only witnessed that on in the first few Shelley epoch transitions (209 through 211)

- It didn't make as big of a difference as I had hoped.

This commit is the result of my attempt to truly minimize the amount of
calculation in TICKF. I've got its run-time down to about 10x slower than a
TICKF that doesn't cross an epoch (for 209 through 2011), but there are
significant subtleties to consider, which the comments in this commit only
partly discuss.

***This is an exploratory Proof-Of-Concept only.***