This commit finalizes the few previous ones that do the ShelleyMA split:
This PR splits the package `cardano-ledger-shelley-ma` into two
packages that implement proper eras:

* `cardano-ledger-allegra`
* and `cardano-ledger-mary`.

There is no longer a `ShelleyMAEra ma c`, but proper types that replace it:
* `AllegraEra c`
*  and `MaryEra c`.

Conceptually the only thing that is different between the two eras is the
support of multi assets by the latter era, therefore the only type that
had seen a somewhat impactful change was the `TxBody`.

There is no longer `MATxBody`, but two:
* `AllegraTxBody`
*  and `MaryTxBody`.

Care was take to reduce duplication so, both of those types are backed
by the same raw type underneath. This also means that corresponding
type class was split as well:

There is no longer `ShelleyMAEraTxBody`, but instead these two:
* `AllegraEraTxBody`
* and `MaryEraTxBody`.