memoize stake pool stake distribution in snapshots
Calculating the stake pool distribution in the NEWEPOCH rule is
problematic due to its role in the consensus layer. When a node checks
to see if it is the leader for a slot crossing the epoch boundary, the
node re-computes the distribution every second, and the
calculation takes over a second. As a stop gap measure, we memoize the
computation by creating a thunk for it at the moment when all the data
is available, which happens to be an entire epoch before it is needed.
In particular, we create the thunk inside the `SnapShots` record in the
SNAP rule.