Added new rules to Conway
Made `PPUPState` and `PPUPPredFailure` into type families so that we can
switch out the concrete instances in Conway and future eras.

Added `EraTallyState` type family so that we don't have to make
`LedgerState` into a type family and we can just set the tally state to
a unit in pre-Conway eras.