Translate bad `Ptr`s to valid `Ptr`s during translation to Conway.
Bad pointers and ugly addresses are not allowed in transactions starting
with Babbage era. However, Ptr's that did land on chain in prior eras are
still present in the LedgerState. Thanks to deprecation of `Ptr`s we can
simplify the Ptr decoding logic, but only after the bad ones have been
translated.