Fixing the Mint field translation.
It was incorrectly fixed in 21825f12729ab6ab97097fa9aa95ba4a5a15672d

This is due to the fact that the Map in Plutus is not like a Map in
Haskell's containers library. Plutus Map is an assoc Map that is based
on lists and the order of elements actually matters. So, the original
fix placed the ADA value at the end, while the original and correct
implementation, as in this commit, places it at the beginning.