Switch from using Voting role at the value level to the type level.
This commit fixes a number of issues related to the `VotingProcedure`.

First and foremost the `Voting` key role is dedicated for DReps only,
but it was being used for Stake Pools and Consitutional Committee
members as well. Using the appropriate key roles makes tarcking of
VoterRole at the value level, not only redundant, but also wrong.

Votes where tracked in a single Map with `(VoterRole, Credential 'Voting c)`
as a key, which was inefficient and wrong at the same type. It was wrong
because it required key role casting, which actually led to a bug of
checking StakePool ids against Stake Credential Map.

Counting up the StakePool votes now uses the Stake Pool Distribution Map
instead of Stake Credential, which fixes the bug that was mentioned above.

ScriptHashes are no longer allowed for StakePool votes. This might be
reverted in the future, but hopefully not. See #3512 for more discussion
on this issue.

Fix CDDL for changes to the new VotingProcedure field in the TxBody.