Improve clarity and performance of collateral Non-ADA validation:
Predicate checks are lazy in their argument, because it is possible to
run STS rules without doing any validation. Current Non-Ada validation
ignored this capability and forced validation regardless of the options
provided to the STS runner.

This commit also improves the clarity on the implementation of
validation, by explaining all possible cases.