Add a set of tests that check upgradeability
The idea behind tests is to serialize a type in the previous era and
deserialize it in the current era, while checking that upgrading
functionality produces the same result.

Noteable type family that does not translate through serialization is
the Script on the Allegra/Alonzo boundary, hence it is treated
specially in tests.