Increase the threshold for StakePools and adjust formula
Previous formula treaded `no` and `abstain` votes equivalently: `t = y / s`

Current formula accounts for `abstain` votes: `t = y / (s - a)`, which
is the same as: `t = (y / s) / (1 - a / s)`. Latter version is used for
efficiency reasons.

Temoprarely threashold is hard coded to 51% for StakePools for SanchoNet.
This will be changed to the actual `ppPoolVotingThresholdsL` Protocol
Parameter in the future.

Fixup the test, to account for the threshold change.