Add vsNumDormantEpochs field to VState.
This field counts the number of contiguous epochs in which the number
of active governance proposals that can be voted on, remained zero. It
will be incremented in every EPOCH if the number of active governance
actions that can be voted on, remains zero. It will be reset when a new
governance proposal is submitted successfully.