Enhance `CommitteeMembersState` query to return quorum and NoConfidence:
When designing the query we forgot about two important pieces of
information relevant for the constitutional commitee, namely the quorum
value and the overall presence of the committee. There is a difference
between an empty committee and no committe, the latter means that the
system is in a state of no confidence.