Bring the Plutus Language version to the type level
This is a major change to how we represent Plutus Scripts. This affects
how we collect and evaluate plutus scripts. There is now a unified
interface for all plutus versions and incorrect usage of an unsupported
plutus version in all eras is prevented by the type system.

Translation of PlutusContext has also been overhauled. There is still
quite a bit duplication happens with TxInfo translation, but this will
be taken care of by a subsequent commit.