Introduce PlutusPurpose:
ScriptPurpuse and RdmrPtr are very much related concepts. This commit
replaces them with a single type family that is parametric on the era
and on the type that does the selection whether it is a plutus script
purpose or it is a redeemer pointer. Those types are called `AsItem` and
`AsIndex` respectively.

Besides making plutus script purpose parametric this commit also
introduces new purposes for Conway: `Voting` and `Proposing`