Remove `sorted_cost_model_keys.txt` as the reference:
Reference for order paramter names should be derived from Plutus,
rather than listed in the file in the ledger repo. Starting with Conway
cost model parameters decoding is lenient, in that it accepts even
unknown parameters, so there is no point in listing all of the ones that
are known.