Disable generation of AlonzoPlutusPurpose. See #4110
Unfortunately we can't fix this problem for Babbage era, because that
would break the node-to-client protocol. So, we need to wait until we
are in Conway to fix this serialization issue with `PlutusPurpose`. With
respect to fixing this problem for Conway, this is already done. It has
been fixed with introduction of `ConwayPlutusPurpose` type in #3960