Fix `estimateMinFeeTx w/ bootstrap` test
In the test, the overestimations were not scaled by the linear factor
`a` of the fee (formula for min fee calculation: `a * txSize tx + b`).
Only the overestimated transaction size was calculated
and it was added to the fee without scaling. This caused the tests to
fail when the linear factor was zero.