Use the Cabal-syntax package instead of the Cabal package
The imports we use in Cardano.Chain.Update.SystemTag were moved to
Cabal-syntax in version 3.8. Depending on Cabal causes an old version to
be used instead (3.2). It's not clear to me how much longer this will
continue to work, and I did in any case have problems building ledger
from the cardano-node nix shell.