Fix a problem where HFC is out of sync with ledger
There is a bug where PPUP rule would not update the PParams,
while consensus Hard Fork Combinator (HFC) would still think that the
update was successful.

This degenerate case could have only happened whenever there was a
ProposedPPUpdates that contained an update to major protocol version
that is expected to be for the new era as well as contained an invalid
update to at least one of the `ppMaxTxSizeL`, `ppMaxBHSizeL` or `ppMaxBBSizeL`
parameters.