DRepDistr: iterate over UMap in chunks.
The DRepDistr was being calculated by iterating over the StakeDistr,
which meant that delegations that have only rewards and no UTxOs would
be left out. This commit changes the calculation to iterate over the
UMap in chunks instead, adding the corresponding rewards and StakeDistr
in the calculation and producing a DRepDistr for all DReps that are
either registered or one of the eternal ones, irrespective of whether
they have an associated StakeDistr.

Also add `UMap.umElemDRepDelegatedReward` to get only those rewards that
have an active DRep delegation.