Change how we report tokens burned in `ValueNotConservedUTxO`:
`ValueNotConservedUTxO` cannot handle negative values, therefore
we need to fix how this failure is being reported. Instead of
reporting negative values for burned tokens in consumed, we
change it to report it as positive in the produced.

In order to implement above fix optional mapping, filtering
and pruning for MultiAsset has been added as well.