Fix strange CI failure.
The range of generated expired epochs and unexpired epochs overlaped at
point 100, which caused the `genExpiredOrResigned` to assume that all
`Yes` voting committee members are to be expired, but according to the
spec we consider members expired only after the current epoch is greater
than their expiry epoch, whereas in this case both were 100.

The range of the generator has been adjusted according to these
semantics.