Change zero valued testing cost models
Plutus has switched to variable number of parameters approach and
increates the number of parameters for PlutusV2. Moreover testing cost
models on the plutus side have not cought up yet.

In any case, until we are in conway we can't vary number of parameters
on the ledger side for PlutusV1/PlutusV2, thereforethus the switch to a
hardcoded number of cost model parameters.

Temporarely make test cost model contain all zeros.