GHA: Work around a bug in actions/download-artifact
We recently started getting "Unable to download and extract artifact"
errors. Turning on debug logging reveals that the error is during
extraction, and the message is "Invalid signature in zip file".

Downloading artifacts from the URL given in the upload step show that
the 8.10.7 archive can't be extracted using the `unzip-stream` library
that `actions/download-artifact` uses. Debug-logging that operation gives
"Unexpected signature in zip file: 0x16d4b50 "PKm", skipped 1 bytes"

Compressing the tar file seems to make a difference, although it's not
clear why, so this probably isn't a long-term solution.