Added minitracing over things with ExecSpecRule instances. Add a few missing constraints from CERT and GOVCERT.
Removed screening out some Signals. They were screened out because of bugs in
the Agda MAlonzo code, or because the conformance tests could not handle
differences between the Spec and the Implementation. An example of this was
screening out the RetirePool PoolCert because the implementation and the spec
changed the state in different rules. This caused all sorts of problems by
breaking the generators. So now all signals are generated fairly, and and
broken conformance tests are moved from 'prop' to 'xprop' until the MAlonzo
code is fixed, or differences can be adjusted.