Generate Shelley CDDL using Huddle.
- Switch existing tests to using the generated CDDL
- Add additional Huddle based tests for Shelley types
- Add a tool to regenerate the Shelley CDDL from Huddle
- Move additional core types to the core CDDL
- Make a few fixes in the Shelley Huddle spec

There is one unusual thing here: the size bound on
the max block header size in the protocol param
update. This does not reflect the original CDDL,
but it is consistent with the FromCBOR instance
and the underlying data type in PParams. I can
only assume that the CDDL generator wasn't
exploring the whole range and thus never found
this error.