Disbale generation of a Ptr map for IncrementalStake
Tomorrow we are going into Conway, so this Map will no longer be useful
as soon as we clean up the Ptr resolution. If we generate Ptr map in
IncrementalStake we will run into issues with roundtripping
serialization for Conway. So the easiest path forward is to disable
generation now and completely get rid of the field later.