Provide and use a helper function to disable treasury expansion
This fixes some tests that broke due to the shortening of the stability
window to a more realistic value (less than the epoch length). Fees
were increasing the treasury and so were throwing off calculations of
the expected treasury value. Previously, these fees didn't reach the
treasury because the end of their stability window hadn't arrived.