Make sure that the fix for #4772 is done safely
We can no longer fix this issue without a hardfork. The danger of fixing
it without a hardfork would be that discrepencies in the ledger state
could lead to different nodes having different DRep delegations when we
out of the bootstrap, if such bug to be manifested on mainnet.

This commit ensures that the fix is applied during the hardfork out of
the bootstrap phase.