Drop dependency on `cardano-prelude`
Reflect usage of `unsafeIndex` with the bounds of `bytestring`.

`bytestring-0.11.3.0` was accidentally released with `unsafeIndex`
hidden from the export list. This issues was promptly fixed in
`bytestring-0.11.4.0`. So, we just need to exclude the affected version
in cabal file instead of rely on some custom implementation of the function
in a package that we don't normally use in ledger.