Make `FlatTerm` part of roundtrip tests for `MemoBytes` types succeed
This is a bit counterintuitive, but it is not possible to decode from
`FlatTerm`, while retriving the original bytes. That is because upon
conversion from `FlatTerm` there are no bytes available to the `cborg`
`Decoder`. We can disable this part of roundtrip test if we ignore the
failure of missing bytes.

See this comment for more info:
https://github.com/well-typed/cborg/blob/f702f0a0485c5c00fba1244949640faa6a59b5f0/cborg/src/Codec/CBOR/FlatTerm.hs#L468-L475