Implement "Multiple identical certificates" Alonzo UTxOW test
The transaction used in this test succeeds in Conway and fails in Alonzo
and Babbage.

The old test suite that this was converted from didn't run all the
rules, only the UTXOW ones. Although this transaction passes the UTXOW
rules it fails DELEG in Alonzo and Babbage, so we can't avoid it being
a failing transaction there. However, by checking that it has only a
DELEG predicate failure we can be sure that it did pass the UTXOW
rules.

Conway's DELEG rules handle the multiple certificates more gracefully,
so this test doesn't fail from Conway onwards.