Implement "Multiple identical certificates" Alonzo UTxOW test
It was necessary to split it between the valid and invalid tests because
it succeeds in Conway and fails in Alonzo and Babbage.

The old tests don't run all the rules, only the UTXOW ones. Although
this test passes the UTXOW rules it fails others in Alonzo and Babbage,
so we can't avoid it being a failing test there. However, by checking
that it has only a DELEG predicate failure we can be sure that it did
pass the UTXOW rules.

Conway's DELEG rules handle the multiple certificates more gracefully,
so this test doesn't fail from Conway onwards.