Untaint era definitions, while preserving the safety warning
Marking individual types with a warning is definitely not what we want
to do, because it will affect all of the downstream users and will
result in a warning until they manually disable it.

We need to mark the module with the warning, instead of the types,
because types are safe to use, unless they are imported from the
internal module.

In order for this to work, we need to define the types in a hidden
module first and then re-export them from an internal module that is
marked with a custom warning.