Fix `FromJSON` instance for `CostModels`
The problem lies in the fact that starting with Conway era number of cost
model can vary, while number of parameters in any genesis file must stay
the same, since that was the number of parameters with which era was
initiated with.

This PR:
* Fixes the parsing where addition of new cost model parameters in a newer
version of plutus results in a failure, unless new parameters are added to
the genesis file, which would be a wrong thing to do.
* Fixes the total number of parameters with which Conway era has started
with. This was a not really a problem, since parsing for cost model
parameters in Conway did not enforce the initial number
* Start enforcing the initial number of parameters in the Conway Genesis