Improve perfomance and reduce overhead of SnapShot creation:
There was some duplicate computation as well as redundant retention of
unnecessary data in the `SnapShots`. This commit introdcues a better
approach that eliminates the need for the above overhead, while
temporarely keeping the old appraoch with some assertions that will be
further used to confirm accuracy on mainnet data. Keeping old approach
should not incur any overhead, since it was made lazy and is only forced
during testing with assertions. The plan is to remove this old approach
before the next release, however there is no danger to go to production
with these follow up changes, due to lazy evaluation of an old approach.

See my comments on #5516 for more details on individual optimizations.