Make ticking a bit more realistic in Imp tests
Current implementation of ticking in ImpSpec is a bit naive, since it
calls `TICK` rule for every slot. This is not only slow, but it also
does not match the reality. On an actual running node TICK is called
on every block, which happens on average every `1/activeSlotCoefficient`
number of slots. On mainnet this is about every 20 slots or every 20
seconds.

Switching to this approach makes imp tests significantly faster, namely
about 80% faster on the example of Conway imp tests.